apps/bkt-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-firestore==2.20.0
pydantic==2.10.3
structlog==24.4.0
httpx==0.28.1
